feat(raw data): 1-convertDataFromDatToCsv, 2-readDataFromCsv and update ratings.csv
